Extract repackaging utility functions
